+++ CHANGES.rst
@@ -5,6 +5,13 @@ Hiveeyes MicroPython Datalogger CHANGES
 
 Development
 ===========
+
+
+2019-06-17 0.4.0
+================
+
+**Getting real.**
+
 - Upgrade to ``Pycom MicroPython 1.20.0.rc11``
 - Stop leaking sensible information into settings output
 - Improve documentation